+++ example/lib/pages/header_page.dart
@@ -1,5 +1,5 @@
-import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:flutter/material.dart';
+import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 
 import '../native_app.dart';
 
@@ -10,19 +10,19 @@ class HeaderPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return ListView(
       children: <Widget>[
-        const Padding(padding: EdgeInsets.all(sbbDefaultSpacing), child: ThemeModeSegmentedButton()),
+        const Padding(padding: EdgeInsets.all(SBBSpacing.medium), child: ThemeModeSegmentedButton()),
         const SBBListHeader('No leading Widget'),
         const SBBHeader(title: 'Title', automaticallyImplyLeading: false),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         const SBBListHeader('Menu'),
         const SBBHeader.menu(title: 'Title'),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         const SBBListHeader('Back'),
         const SBBHeader.back(title: 'Title'),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         const SBBListHeader('Close'),
         const SBBHeader.close(title: 'Title'),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         const SBBListHeader('With Callback on SBB Signet'),
         SBBHeader(
           title: 'Title',
@@ -30,7 +30,7 @@ class HeaderPage extends StatelessWidget {
           onPressedLogo: () => Navigator.maybePop(context),
           logoTooltip: 'Back to home',
         ),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
       ],
     );
   }