+++ example/lib/pages/list_item_page.dart
@@ -17,10 +17,7 @@ class ListItemPage extends StatelessWidget {
       slivers: [
         SBBSliverHeaderbox.custom(child: ThemeModeSegmentedButton()),
         SliverPadding(
-          padding: const EdgeInsets.symmetric(
-            vertical: sbbDefaultSpacing * .5,
-            horizontal: sbbDefaultSpacing * .5,
-          ).copyWith(bottom: sbbDefaultSpacing * 3),
+          padding: const EdgeInsets.all(SBBSpacing.xSmall),
           sliver: SliverList.list(
             children: [
               SBBListHeader('Listed'),
@@ -89,10 +86,10 @@ class ListItemPage extends StatelessWidget {
                   ).toList(),
                 ),
               ),
-              SizedBox(height: sbbDefaultSpacing),
+              SizedBox(height: SBBSpacing.medium),
               SBBListHeader('Boxed'),
               Column(
-                spacing: sbbDefaultSpacing * .5,
+                spacing: SBBSpacing.xSmall,
                 children: [
                   SBBListItemBoxed(
                     titleText: 'Default',