+++ lib/src/button/theme/sbb_button_style_x.dart
@@ -27,7 +27,7 @@ extension SBBButtonStyleX on SBBButtonStyle {
   ButtonStyle get baseButtonStyle => ButtonStyle(
     overlayColor: WidgetStatePropertyAll<Color>(SBBColors.transparent),
     fixedSize: WidgetStatePropertyAll<Size>(const Size.fromHeight(SBBInternal.defaultButtonHeight)),
-    padding: WidgetStatePropertyAll<EdgeInsets>(const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing)),
+    padding: WidgetStatePropertyAll<EdgeInsets>(const EdgeInsets.symmetric(horizontal: SBBSpacing.medium)),
     elevation: WidgetStatePropertyAll<double>(0),
     tapTargetSize: MaterialTapTargetSize.shrinkWrap,
     mouseCursor: WidgetStateMouseCursor.clickable,