+++ lib/src/picker/sbb_date_picker.dart
@@ -83,7 +83,7 @@ class SBBDatePicker extends StatefulWidget {
         mainAxisSize: MainAxisSize.min,
         children: [
           Padding(
-            padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
+            padding: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
             child: SBBContentBox(
               child: SBBDatePicker(
                 initialDate: modalDate,
@@ -99,7 +99,7 @@ class SBBDatePicker extends StatefulWidget {
             ),
           ),
           Padding(
-            padding: const EdgeInsets.all(sbbDefaultSpacing),
+            padding: const EdgeInsets.all(SBBSpacing.medium),
             child: ListenableBuilder(
               listenable: selectedButtonEnabled,
               builder: (context, _) {