+++ test/sbb_pagination_test.dart
@@ -9,24 +9,24 @@ void main() {
     final widget = Column(
       children: [
         const SBBPagination(numberPages: 1, currentPage: 0),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         const SBBPagination(numberPages: 2, currentPage: 0),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         const SBBPagination(numberPages: 5, currentPage: 4),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         // floating
         const SBBPagination(numberPages: 1, currentPage: 0, isFloating: true),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         // add containter to see the shadow and floating box
         Container(
           color: SBBColors.white,
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
+          padding: const EdgeInsets.all(SBBSpacing.medium),
           child: const SBBPagination(numberPages: 2, currentPage: 0, isFloating: true),
         ),
-        const SizedBox(height: sbbDefaultSpacing),
+        const SizedBox(height: SBBSpacing.medium),
         Container(
           color: SBBColors.white,
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
+          padding: const EdgeInsets.all(SBBSpacing.medium),
           child: const SBBPagination(numberPages: 5, currentPage: 4, isFloating: true),
         ),
       ],